feat(demo-server): add endpoint to retrieve signed upload url - EUBFR-43
