+++ config.example.json
@@ -1,5 +1,10 @@
 {
   "eubfr_env": "dev",
   "region": "eu-central-1",
-  "stage": "uuid"
+  "stage": "uuid",
+  "demo": {
+    "AWS_ACCESS_KEY_ID": "",
+    "AWS_SECRET_ACCESS_KEY": "",
+    "SIGNED_UPLOADS_API_ID": ""
+  }
 }
