+++ demo/package.json
@@ -21,6 +21,6 @@
     "webpack": "3.6.0"
   },
   "dependencies": {
-    "uuid": "3.1.0"
+    "aws4": "1.6.0"
   }
 }
