--- demo/server/src/lib/extractUsername.js
@@ -1,4 +1,0 @@
-export const extractUsername = userArn =>
-  userArn ? userArn.split(':')[5].replace('user/', '') : '';
-
-export default extractUsername;
